Split fdb tests into 2 files
